adjust double sidebar

+++ src/resources/views/components/sidebar/header.blade.php
@@ -1,6 +1,6 @@
 @props([
     'class' => '',
-    'variant' => 'double',
+    'variant' => 'single',
 
 ])
 @php
